pyproject.toml
[project]
name = "langgraph-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "ipykernel>=7.2.0",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "langchain-openai>=1.1.10",
    "langchain-tavily>=0.2.17",
    "langgraph>=1.0.10",
    "langgraph-cli[inmem]>=0.4.14",
    "langsmith>=0.7.9",
    "python-dotenv>=1.2.1",
]
